# internal/api/handlers/camera.go
package handlers

import (
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
)

// CameraHandler handles camera-related endpoints
type CameraHandler struct{}

// NewCameraHandler creates a new camera handler
func NewCameraHandler() *CameraHandler {
	return &CameraHandler{}
}

// @Summary Add camera
// @Description Add a new camera to the worker for processing
// @Tags camera
// @Accept json
// @Produce json
// @Param camera body object true "Camera configuration"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /cameras [post]
func (h *CameraHandler) AddCamera(c *gin.Context) {
	var request struct {
		ID      string `json:"id" binding:"required"`
		RTSPUrl string `json:"rtsp_url" binding:"required"`
		Name    string `json:"name"`
		Enabled bool   `json:"enabled"`
	}

	if err := c.ShouldBindJSON(&request); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error":   "Invalid request format",
			"details": err.Error(),
		})
		return
	}

	// TODO: Add actual camera management logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"camera_id": request.ID,
		"message":   "Camera added successfully",
		"timestamp": time.Now().Unix(),
	})
}

// @Summary Remove camera
// @Description Remove a camera from the worker
// @Tags camera
// @Accept json
// @Produce json
// @Param id path string true "Camera ID"
// @Success 200 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /cameras/{id} [delete]
func (h *CameraHandler) RemoveCamera(c *gin.Context) {
	cameraID := c.Param("id")

	// TODO: Add actual camera removal logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"camera_id": cameraID,
		"message":   "Camera removed successfully",
		"timestamp": time.Now().Unix(),
	})
}

// @Summary List cameras
// @Description Get list of all cameras and their status
// @Tags camera
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /cameras [get]
func (h *CameraHandler) ListCameras(c *gin.Context) {
	// TODO: Add actual camera listing logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"cameras":   []gin.H{},
		"total":     0,
		"timestamp": time.Now().Unix(),
	})
}

// @Summary Get camera status
// @Description Get status of a specific camera
// @Tags camera
// @Accept json
// @Produce json
// @Param id path string true "Camera ID"
// @Success 200 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /cameras/{id}/status [get]
func (h *CameraHandler) GetCameraStatus(c *gin.Context) {
	cameraID := c.Param("id")

	// TODO: Add actual camera status logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"camera_id": cameraID,
		"status":    "active",
		"timestamp": time.Now().Unix(),
	})
}

// @Summary Get latest frame
// @Description Get the latest processed frame from a camera
// @Tags camera
// @Accept json
// @Produce json
// @Param id path string true "Camera ID"
// @Success 200 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /cameras/{id}/frame [get]
func (h *CameraHandler) GetLatestFrame(c *gin.Context) {
	cameraID := c.Param("id")

	// TODO: Add actual frame retrieval logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"camera_id": cameraID,
		"frame_url": "",
		"timestamp": time.Now().Unix(),
	})
}