# internal/api/handlers/webrtc.go
package handlers

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
)

// WebRTCHandler handles WebRTC-related endpoints
type WebRTCHandler struct{}

// NewWebRTCHandler creates a new WebRTC handler
func NewWebRTCHandler() *WebRTCHandler {
	return &WebRTCHandler{}
}

// @Summary WebRTC control
// @Description Control WebRTC streaming for cameras
// @Tags webrtc
// @Accept json
// @Produce json
// @Param control body object true "WebRTC control"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /webrtc/control [post]
func (h *WebRTCHandler) Control(c *gin.Context) {
	var request struct {
		CameraID string `json:"camera_id" binding:"required"`
		Action   string `json:"action" binding:"required"` // "start" or "stop"
	}

	if err := c.ShouldBindJSON(&request); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error":   "Invalid request format",
			"details": err.Error(),
		})
		return
	}

	// TODO: Add actual WebRTC control logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"camera_id": request.CameraID,
		"action":    request.Action,
		"message":   fmt.Sprintf("WebRTC %s successful", request.Action),
		"timestamp": time.Now().Unix(),
	})
}

// @Summary WebRTC status
// @Description Get WebRTC status for all cameras
// @Tags webrtc
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /webrtc/status [get]
func (h *WebRTCHandler) Status(c *gin.Context) {
	// TODO: Add actual WebRTC status logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"cameras":   []gin.H{},
		"total":     0,
		"timestamp": time.Now().Unix(),
	})
}

// @Summary Get stream URLs
// @Description Get streaming URLs for a specific camera
// @Tags webrtc
// @Accept json
// @Produce json
// @Param id path string true "Camera ID"
// @Success 200 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /webrtc/{id}/urls [get]
func (h *WebRTCHandler) GetStreamURLs(c *gin.Context) {
	cameraID := c.Param("id")

	// TODO: Add actual stream URL logic
	c.JSON(http.StatusOK, gin.H{
		"success":   true,
		"camera_id": cameraID,
		"urls": gin.H{
			"whip": fmt.Sprintf("/whip/%s", cameraID),
			"hls":  fmt.Sprintf("/hls/%s/index.m3u8", cameraID),
		},
		"timestamp": time.Now().Unix(),
	})
}